requirements.txt
flask-socketio
gevent-websocket
flask
flask_cors
python-socketio[asyncio_client]
